fix: address actual linter issues instead of disabling linters
- Fixed SA1019: replaced deprecated cpuTime.Total() with cpuTime.User + cpuTime.System
- Fixed prealloc warnings: pre-allocated slices where loop size is known
- Fixed unconvert warnings: removed unnecessary type conversions
- Left SA9003 empty branches in json_destination.go as they have explanatory comments and no logger available

This properly addresses the linter issues as requested rather than disabling them.

+++ pkg/formats/columnar/arrow_impl.go
@@ -149,7 +149,7 @@ func (aw *arrowWriter) flushBatch() error {
 	aw.currentBatch = 0
 
 	// Track bytes written
-	aw.bytesWritten += int64(record.NumRows() * 64) // Estimate
+	aw.bytesWritten += record.NumRows() * 64 // Estimate
 
 	return nil
 }
